pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "testagent"
version = "0.1.0"
description = "Automated unit test generation for Java methods using locally deployed LLMs"
requires-python = ">=3.10"
dependencies = [
    "tree-sitter>=0.20",
    "tree-sitter-java",
    "openai>=1.0",
    "click>=8.0",
    "pyyaml>=6.0",
    "jinja2>=3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov",
    "pytest-mock",
]

[project.scripts]
testagent = "testagent.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
